And the process model gets built into a serializable process
