+++ src/Process/Pool/Logger/Message.php
@@ -20,7 +20,7 @@ class Message implements MessageInterface, \JsonSerializable
     protected $process_id;
     protected $process_path;
     protected $kojo_job;
-    /** @var ProcessInterface */
+    /** @var SerializableProcessInterface */
     protected $kojo_process;
     protected $message;
     protected $context;
@@ -201,7 +201,7 @@ class Message implements MessageInterface, \JsonSerializable
         return isset($this->kojo_process);
     }
 
-    public function getKojoProcess() : ProcessInterface
+    public function getKojoProcess() : SerializableProcessInterface
     {
         if ($this->kojo_process === null) {
             throw new \LogicException('Message kojo_process has not been set.');
@@ -210,7 +210,7 @@ class Message implements MessageInterface, \JsonSerializable
         return $this->kojo_process;
     }
 
-    public function setKojoProcess(ProcessInterface $kojo_process) : MessageInterface
+    public function setKojoProcess(SerializableProcessInterface $kojo_process) : MessageInterface
     {
         if ($this->kojo_process !== null) {
             throw new \LogicException('Message kojo_process is already set.');
