--- src/Process/Pool/Logger/Message/Process.yml
@@ -1,3 +1,0 @@
-services:
-  neighborhoods.kojo.neighborhoods.kojo.src.process.pool.logger.message.process:
-    class: Neighborhoods\Kojo\Process\Pool\Logger\Message\Process
