--- src/Process/Pool/Logger/Message/Process/Factory.yml
@@ -1,5 +1,0 @@
-services:
-  neighborhoods.kojo.process.pool.logger.message.process.factory:
-    class: Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\Factory
-    calls:
-      - [setProcessPoolLoggerMessageProcess, ['@neighborhoods.kojo.neighborhoods.kojo.src.process.pool.logger.message.process']]
