--- src/Process/Pool/Logger/Message/Process/FromProcessModel.yml
@@ -1,3 +1,0 @@
-services:
-  neighborhoods.kojo.process.pool.logger.message.process.from_process_model:
-    class: Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel
