--- src/Process/Pool/Logger/Message/Process/FromProcessModel/Builder.yml
@@ -1,7 +1,0 @@
-services:
-  neighborhoods.kojo.process.pool.logger.message.process.from_process_model.builder:
-    class: Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel\Builder
-    calls:
-      - [setProcessPoolLoggerMessageProcessFromProcessModelFactory, ['@neighborhoods.kojo.process.pool.logger.message.process.from_process_model.factory']]
-      - [setProcessPoolLoggerMessageProcessFactory, ['@neighborhoods.kojo.process.pool.logger.message.process.factory']]
-
