--- src/Process/Pool/Logger/Message/Process/FromProcessModel/Builder/Factory.yml
@@ -1,5 +1,0 @@
-services:
-  neighborhoods.kojo.process.pool.logger.message.process.from_process_model.builder.factory:
-    class: Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel\Builder\Factory
-    calls:
-      - [setProcessPoolLoggerMessageProcessFromProcessModelBuilder, ['@neighborhoods.kojo.process.pool.logger.message.process.from_process_model.builder']]
