--- src/Process/Pool/Logger/Message/Process/FromProcessModel/Builder/FactoryInterface.php
@@ -1,12 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel\Builder;
-
-use Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel\BuilderInterface;
-
-/** @codeCoverageIgnore */
-interface FactoryInterface
-{
-    public function create() : BuilderInterface;
-}
