--- src/Process/Pool/Logger/Message/Process/FromProcessModel/Factory.yml
@@ -1,5 +1,0 @@
-services:
-  neighborhoods.kojo.process.pool.logger.message.process.from_process_model.factory:
-    class: Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel\Factory
-    calls:
-      - [setProcessPoolLoggerMessageProcessFromProcessModel, ['@neighborhoods.kojo.process.pool.logger.message.process.from_process_model']]
