--- src/Process/Pool/Logger/Message/Process/FromProcessModel/FactoryInterface.php
@@ -1,12 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel;
-
-use Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModelInterface;
-
-/** @codeCoverageIgnore */
-interface FactoryInterface
-{
-    public function create() : FromProcessModelInterface;
-}
