+++ src/Process/Pool/Logger/Message/SerializableProcess.yml
@@ -1,0 +1,3 @@
+services:
+  neighborhoods.kojo.neighborhoods.kojo.src.process.pool.logger.message.serializable_process:
+    class: Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess
