+++ src/Process/Pool/Logger/Message/SerializableProcess/Factory.php
@@ -1,0 +1,17 @@
+<?php
+declare(strict_types=1);
+
+namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess;
+
+use Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcessInterface;
+
+/** @codeCoverageIgnore */
+class Factory implements FactoryInterface
+{
+    use AwareTrait;
+
+    public function create() : SerializableProcessInterface
+    {
+        return clone $this->getProcessPoolLoggerMessageProcess();
+    }
+}
