+++ src/Process/Pool/Logger/Message/SerializableProcess/FactoryInterface.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess;
+
+use Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcessInterface;
+
+/** @codeCoverageIgnore */
+interface FactoryInterface
+{
+    public function create() : SerializableProcessInterface;
+}
