+++ src/Process/Pool/Logger/Message/SerializableProcess/FromProcessModel/Builder/Factory.yml
@@ -1,0 +1,5 @@
+services:
+  neighborhoods.kojo.process.pool.logger.message.serializable_process.from_process_model.builder.factory:
+    class: Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess\FromProcessModel\Builder\Factory
+    calls:
+      - [setProcessPoolLoggerMessageProcessFromProcessModelBuilder, ['@neighborhoods.kojo.process.pool.logger.message.serializable_process.from_process_model.builder']]
