+++ src/Process/Pool/Logger/Message/SerializableProcess/FromProcessModel/BuilderInterface.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel;
+namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess\FromProcessModel;
 
-use Neighborhoods\Kojo\Process\Pool\Logger\Message\ProcessInterface as ProcessInterfaceAlias;
+use Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcessInterface as ProcessInterfaceAlias;
 use Neighborhoods\Kojo\ProcessInterface;
 
 interface BuilderInterface
