+++ src/Process/Pool/Logger/Message/SerializableProcess/FromProcessModel/Factory.php
@@ -1,9 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModel;
+namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess\FromProcessModel;
 
-use Neighborhoods\Kojo\Process\Pool\Logger\Message\Process\FromProcessModelInterface;
+use Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess\FromProcessModelInterface;
 
 /** @codeCoverageIgnore */
 class Factory implements FactoryInterface
