+++ src/Process/Pool/Logger/Message/SerializableProcess/FromProcessModel/FactoryInterface.php
@@ -1,0 +1,12 @@
+<?php
+declare(strict_types=1);
+
+namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess\FromProcessModel;
+
+use Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess\FromProcessModelInterface;
+
+/** @codeCoverageIgnore */
+interface FactoryInterface
+{
+    public function create() : FromProcessModelInterface;
+}
