+++ src/Process/Pool/Logger/Message/SerializableProcess/FromProcessModelInterface.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\Process;
+namespace Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcess;
 
 use Neighborhoods\Kojo\ProcessInterface;
