+++ src/Process/Pool/Logger/MessageInterface.php
@@ -6,7 +6,7 @@ namespace Neighborhoods\Kojo\Process\Pool\Logger;
 
 
 use Neighborhoods\Kojo\Data\JobInterface;
-use Neighborhoods\Kojo\Process\Pool\Logger\Message\ProcessInterface;
+use Neighborhoods\Kojo\Process\Pool\Logger\Message\SerializableProcessInterface;
 
 interface MessageInterface
 {
@@ -30,6 +30,6 @@ interface MessageInterface
 
     public function hasKojoProcess() : bool;
 
-    public function getKojoProcess() : ProcessInterface;
+    public function getKojoProcess() : SerializableProcessInterface;
 
 }
